Correctly recognise when a user is deined access
So far, the "deny" resopnse from an HTTP authentication server was parsed
as string while an atom was expected. As a result, if the server denied
a user it simply happend to work as the plugin would return an error
because of the type mismatch and they user would not be accepted anyway.